Removed V1 Payload support
