+++ sdk/builder.go
@@ -1,4 +1,4 @@
-package v2
+package sdk
 
 import (
 	"fmt"
